Refactor RadioButton and Cards components
This commit removes the RadioButton component and refactors the Cards component.

The RadioButton component has been deleted as it is no longer needed.

In the Cards component, the Card import has been removed, and the component has been updated to use a div instead of a Card.

The styling and structure of the Cards component have been modified. The Card.Img component has been replaced with an img tag, and the Card.ImgOverlay component has been replaced with a div.

The Card.Body component has been replaced with a div, and the Card.Title and Card.Text components have been replaced with h2 and p tags, respectively.

The Button component has been replaced with a button tag.

The propTypes for the Cards component have also been updated.

These changes improve the structure and maintainability of the code.